fix(console): open provider workspace by default
